refactor: decompose popup and navigation menu internals
